Support ipv6 output from ufw status.
The output from ufw status for an ipv6 from any rule adds `(v6)` to the end of
the line
```
$ sudo ufw status
Status: active

To                         Action      From
--                         ------      ----
22/tcp                     ALLOW       Anywhere
22/tcp                     ALLOW       Anywhere (v6)
```

This causes an ipv6 from any rule to be execed on each puppet run.

This commit works out if the ip provided is ipv6 or not and then changes
the regex for the unless.